app/requirements.txt
python-dotenv
flask
flask-jwt-extended
neo4j
openai
google-auth
requests
flask_oauthlib
Werkzeug==2.2.2